requirements.txt
tensorflow
pandas 
dvc
notebook
numpy
matplotlib
seaborn
python-box==6.0.2
pyYAML
tqdm
ensure==1.0.2
joblib
types-PyYAML
scipy
Flask
Flask-Cors
dill==0.3.5.1
from-root==1.0.2
gdown
ultralytics==8.0.28
pytest
pytest-mock
-e .